requirements-docker.txt
chainlit==2.5.5
docx==0.2.4
email-validator==2.1.0.post1
flask==3.0.3
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
pypdf2==3.0.1
python-docx==1.1.2
requests==2.31.0
sqlalchemy==2.0.29
weaviate-client==4.13.2
werkzeug==3.0.1